Update register UI
