+++ src/App.jsx
@@ -1,10 +1,6 @@
-import axios from "axios";
 import "./App.css";
-import { useEffect } from "react";
-import FormLogin from "./component/form.jsx";
 import { LoginPage } from "./page/login/index.jsx";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import { Register } from "./page/register/index.jsx";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import DashBoard from "./component/dashboard/index.jsx";
@@ -19,8 +15,6 @@ import { HomepageLogin } from "./page/homepage/HomepageLogin.jsx";
 import { Hostpage } from "./page/HostPage/Hostpage.jsx";
 import { HostProfile } from "./page/profile/HostProfile.jsx";
 import EditProfileHosts from "./page/party-host/editprofilehosts/index.jsx";
-import ServicePage from "./page/party-host/service/index.jsx";
-import PackagePage from "./page/party-host/package/index.jsx";
 import ReportPage from "./page/party-host/report/index.jsx";
 import ManageAccounts from "./page/dashboad_admin/report_admin/manage_accounts/index.jsx";
 import ReportPageAdmin from "./page/dashboad_admin/report_admin/index.jsx";
@@ -54,7 +48,7 @@ function App() {
     },
     {
       path: "/register",
-      element: <Register />,
+      element: <LoginPage isRegister />,
     },
     {
       path: "/dashboard",
